requirements.txt
PyYAML~=6.0
pendulum~=2.1.2
kubernetes~=19.15.0
packaging~=21.2
setuptools~=58.3.0
apache-airflow>=1.10.2
Flask~=1.1.4
Jinja2~=2.11.3
expandvars~=0.7.0